Fix: More dynamic reading of properties
