+++ source/blender/io/ply/importer/ply_import.cc
@@ -25,7 +25,7 @@
 
 #include "intern/ply_data.hh"
 #include "ply_import.hh"
-#include "ply_import_big_endian.hh"
+#include "ply_import_binary.hh"
 
 namespace blender::io::ply {
